Move cramer_into the test file and use Timo's new linear solver parser to match KanesMethod.
