+++ sympy/physics/mechanics/tests/test_kane3.py
@@ -1,3 +1,4 @@
+from sympy import cacheit, zeros, ImmutableMatrix
 from sympy.core.numbers import pi
 from sympy.core.symbol import symbols
 from sympy.functions.elementary.miscellaneous import sqrt
@@ -6,8 +7,34 @@ from sympy.matrices.dense import Matrix
 from sympy.physics.mechanics import (ReferenceFrame, dynamicsymbols,
                                      KanesMethod, inertia, Point, RigidBody,
                                      dot)
-from sympy.physics.mechanics.functions import cramer_solve
 from sympy.testing.pytest import slow, ON_CI, skip
+
+
+# TODO : This is a temporary implementation and can be removed once we support
+# a new matrix solver A.CramerSolve(b).
+@cacheit
+def det_laplace(matrix):
+    n = matrix.shape[0]
+    if n == 1:
+        return matrix[0]
+    elif n == 2:
+        return matrix[0, 0] * matrix[1, 1] - matrix[0, 1] * matrix[1, 0]
+    else:
+        return sum((-1) ** i * matrix[0, i] *
+                   det_laplace(matrix.minor_submatrix(0, i)) for i in range(n))
+
+
+def cramer_solve(A, b, det_method=det_laplace):
+    def entry(i, j):
+        return b[i, sol] if j == col else A[i, j]
+
+    A_imu = ImmutableMatrix(A)  # Convert to immutable for cache purposes
+    det_A = det_method(A_imu)
+    x = zeros(*b.shape)
+    for sol in range(b.shape[1]):
+        for col in range(b.shape[0]):
+            x[col, sol] = det_method(ImmutableMatrix(*A.shape, entry)) / det_A
+    return x
 
 
 @slow
